Math.polynomial {polynom}	R Documentation

Math Group Methods for Polynomials

Description

Group method for functions in the Math group.

Usage

## S3 method for class 'polynomial'
Math(x, ...)

Arguments

x
an object of class "polynomial".

...
further arguments to be passed to or from methods, such as digits.

Details

Most math group functions are disallowed with polynomial arguments. The only exceptions are ceiling, floor, round, trunc, and signif which may be used to transform the coefficients accordingly.

Value

A polynomial with transformed coefficients.

See Also

Ops.polynomial, Summary.polynomial.

Examples

op <- options(digits=18)
p <- poly.from.values(1:4, (2:5)^2)
## 1 + 2.00000000000001*x + x^2
p <- round(p)
## 1 + 2*x + x^2
options(op)

[Package polynom version 1.4-1 Index]