Ops.polynomial {polynom}	R Documentation

Arithmetic Ops Group Methods for Polynomials

Description

Allows arithmetic operators to be used for polynomial calculations, such as addition, multiplication, division, etc.

Usage

## S3 method for class 'polynomial'
Ops(e1, e2)

Arguments

e1
an object of class "polynomial".

e2
an object of class "polynomial".

Value

A polynomial got by performing the operation on the two arguments.

See Also

Math.polynomial, Summary.polynomial.

Examples

p <- polynomial(c(1, 2, 1))
## 1 + 2*x + x^2
r <- poly.calc(-1 : 1)
## -1*x + x^3
(r - 2 * p)^2
## 4 + 20*x + 33*x^2 + 16*x^3 - 6*x^4 - 4*x^5 + x^6

[Package polynom version 1.4-1 Index]