as.function.polynomial {polynom}	R Documentation

Coerce a Polynomial to a Function

Description

Takes a polynomial argument and constructs an R function to evaluate it at arbitrary points.

Usage


## S3 method for class 'polynomial'
as.function(x, ...)

Arguments

x
An object of class "polynomial".

...
further arguments to be passed to or from methods.

Details

This is a method for the generic function as.function.

The polynomial is evaluated within the function using the Horner scheme.

Note that you can use the model-oriented predict method for polynomials for purpose of evaluation (without explicit coercion to a function), see the example below.

Value

A function to evaluate the polynomial p.

See Also

as.function, predict.polynomial

Examples


pr <- (poly.calc(-1:1) - 2 * polynomial(c(1, 2, 1)))^2
pr
## 4 + 20*x + 33*x^2 + 16*x^3 - 6*x^4 - 4*x^5 + x^6
prf <- as.function(pr)
prf
## function (x) 
## 4 + x * (20 + x * (33 + x * (16 + x * (-6 + x * (-4 + x * (1))))))
## <environment: 0x402440f0>
prf(-3:3)
##  1024 64  0 4 64 144 64
predict(pr, -3:3)
##  1024 64  0 4 64 144 64

[Package polynom version 1.4-0 Index]