deriv.polynomial {polynom}	R Documentation

Differentiate a Polynomial

Description

Calculates the derivative of a univariate polynomial.

Usage


## S3 method for class 'polynomial'
deriv(expr, ...)

Arguments

expr
an object of class "polynomial".

...
further arguments to be passed to or from methods.

Details

This is a method for the generic function deriv.

Value

Derivative of the polynomial.

See Also

integral.polynomial, deriv.

Examples


pr <- poly.calc(1:5)
pr
## -120 + 274*x - 225*x^2 + 85*x^3 - 15*x^4 + x^5
deriv(pr)
##  274 - 450*x + 255*x^2 - 60*x^3 + 5*x^4

[Package polynom version 1.4-0 Index]