lines.polynomial {polynom}	R Documentation

Lines Method for Polynomials

Description

Add a polynomial to an existing plot usually as a line plot.

Usage


## S3 method for class 'polynomial'
lines(x, len = 1000, xlim = NULL, ylim = NULL, ...)

Arguments

x
an object of class "polynomial".

len
size of vector at which evaluations are to be made.

xlim, ylim
the range of x and y values with sensible defaults.

...
additional arguments as for the lines generic.

Details

This is a method for the generic function lines.

Lines representing the given polynomial are added to an existing plot. Values outside the current plot region are not shown.

See Also

lines, points, points.polynomial, plot, plot.polynomial.

Examples


plot  (poly.calc(-1:5))
lines (poly.calc( 2:4), lty = 2)
points(poly.calc(-2:6), pch = 4)

[Package polynom version 1.4-0 Index]