monic {polynom}	R Documentation

Monic Polynomials

Description

Convert a polynomial to monic form by dividing by the leading coefficient.

Usage


monic(p)

Arguments

p
A polynomial. A warning is issued if the polynomial is identically zero.

Details

Similar in effect to p/as.numeric(p[length(p)]) but with some safeguards against leading zero coefficients.

Value

A polynomial proportional to p with leading coefficient 1.

See Also

change.origin

[Package polynom version 1.4-0 Index]