plot.polynomial {polynom}	R Documentation

Plot Method for Polynomials

Description

Plots polynomials, optionally allowing the “interesting” region to be automatically determined.

Usage


## S3 method for class 'polynomial'
plot(x, xlim = 0:1, ylim = range(Px), type = "l",
                len = 1000, ..., xlab = "x", ylab = "P(x)")

Arguments

x
an object of class "polynomial".

xlim
the range to be encompassed by the x axis.

ylim
the range to be encompassed by the y axis.

type
as for plot().

len
number of x points drawn.

...
additional arguments as for plot.

xlab,ylab
graphical parameters.

Details

This is a method for the generic function plot.

A plot of the polynomial is produced on the currently active device. Unless otherwise specified, the domain is chosen to enclose the real parts of all zeros, stationary points and zero itself.

See Also

plot, lines, points, lines.polynomial, points.polynomial.

Examples


plot(p <- poly.calc(-1:5))

[Package polynom version 1.4-0 Index]