points.polynomial {polynom}	R Documentation

Points Method for Polynomials

Description

Add a polynomial to an existing plot usually as a point plot.

Usage


## S3 method for class 'polynomial'
points(x, length = 100, ...)

Arguments

x
an object of class "polynomial".

length
size of x vector at which evaluations are to be made.

...
additional arguments as for the points generic.

Details

This is a method for the generic function points.

Points representing the given polynomial are added to an existing plot. Values outside the current plot region are not shown.

See Also

plot, lines, points, plot.polynomial, lines.polynomial.

Examples


plot(poly.calc(-1:5))
lines(poly.calc(2:4), lty=2)
points(poly.calc(-2:6), pch=4)

[Package polynom version 1.4-0 Index]