predict.polynomial {polynom}	R Documentation

Evaluate a Polynomial

Description

Evaluate a polynomial at a given numeric or polynomial argument.

Usage

## S3 method for class 'polynomial'
predict(object, newdata, ...)

Arguments

object
A polynomial object to be evaluated.

newdata
Argument at which evaluation is requested. May be numeric or itself a polynomial

...
Not used by this method.

Details

This is a method for the generic function predict.

The polynomial is evaluated according to the Horner scheme for speed and numerical accuracy.

Value

Evaluated object of the same class as newdata.

See Also

as.function.polynomial

[Package polynom version 1.4-1 Index]