zSummary.polynomial {polynom}	R Documentation

Summary Group Methods for Polynomials

Description

Allows summary group generics to be used on polynomial arguments.

Usage

## S3 method for class 'polynomial'
Summary(..., na.rm = FALSE)

Arguments

...
R objects, the first supplied of class "polynomial".

na.rm
logical: should missing values be removed?

Details

For the sum and prod functions, the sum and product of the given polynomials, respectively. For the other members of the Summary group, an error is returned.

Note that one could order polynomials by divisibility, and define min and max as the corresponding lattice meet and join, i.e., the greatest common divisor and the least common multiple, respectively. This is currently not provided: instead, functions GCD and LCM should be called directly.

See Also

Math.polynomial, Ops.polynomial

[Package polynom version 1.4-1 Index]